Renamed TestClassUtils to TestUtils, moved report expectation method there
